fix(shell): address mobile drawer review feedback
