+++ src/components/app/AppShellHeader.tsx
@@ -2,8 +2,8 @@ import { Link } from "@tanstack/react-router";
 import { BookOpenText, UserRound } from "lucide-react";
 
 import { Button } from "@/components/ui/button";
-import { getProfileAvatarFallbackLabel } from "@/lib/profilePresentation";
 
+import { HeaderAvatar } from "./HeaderAvatar";
 import { MobileNavDrawer } from "./MobileNavDrawer";
 
 import type { AuthActionState, HeaderNavItem } from "./AppShellHeader.types";
@@ -97,32 +97,5 @@ export function AppShellHeader({
         <MobileNavDrawer authAction={authAction} navItems={navItems} />
       </div>
     </header>
-  );
-}
-
-type HeaderAvatarProps = {
-  avatarUrl: string | null;
-  label: string;
-};
-
-function HeaderAvatar({ avatarUrl, label }: HeaderAvatarProps): JSX.Element {
-  if (avatarUrl !== null) {
-    return (
-      <img
-        alt=""
-        aria-hidden="true"
-        className="size-5 rounded-full border border-border object-cover"
-        src={avatarUrl}
-      />
-    );
-  }
-
-  return (
-    <span
-      aria-hidden="true"
-      className="inline-flex size-5 items-center justify-center rounded-full border border-border bg-muted text-[0.65rem] font-semibold text-foreground"
-    >
-      {getProfileAvatarFallbackLabel(label)}
-    </span>
   );
 }
